Move AnvilScaleJujuStep to anvil.commands to avoid circular import. Add custom JujuManifest to remove scaling_args. Add is_skip override to determine whether it needs to be run and determine which machines need controllers
