+++ anvil-python/anvil/commands/juju.py
@@ -13,12 +13,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import json
+import logging
 from os import environ
 import os.path
 import subprocess
 
 from rich.status import Status
+from sunbeam.commands.juju import ScaleJujuStep
 from sunbeam.jobs.common import BaseStep, Result, ResultType
+
+from anvil.utils import machines_missing_juju_controllers
+
+LOG = logging.getLogger(__name__)
 
 
 class JujuAddSSHKeyStep(BaseStep):
@@ -54,3 +61,59 @@ class JujuAddSSHKeyStep(BaseStep):
                 message="Could not find public ssh key (~/.ssh/id_rsa.pub)",
             )
         return Result(ResultType.COMPLETED)
+
+
+class AnvilScaleJujuStep(ScaleJujuStep):
+    def run(self, status: Status | None = None) -> Result:
+        cmd = [
+            self._get_juju_binary(),
+            "enable-ha",
+            "-n",
+            str(self.n),
+            *self.extra_args,
+        ]
+        LOG.debug(f'Running command {" ".join(cmd)}')
+        process = subprocess.run(
+            cmd, capture_output=True, text=True, check=True
+        )
+        LOG.debug(
+            f"Command finished. stdout={process.stdout}, stderr={process.stderr}"
+        )
+        cmd = [
+            self._get_juju_binary(),
+            "wait-for",
+            "application",
+            "-m",
+            "admin/controller",
+            "controller",
+            "--timeout",
+            "15m",
+        ]
+        self.update_status(status, "scaling controller")
+        LOG.debug("Waiting for HA to be enabled")
+        LOG.debug(f'Running command {" ".join(cmd)}')
+        process = subprocess.run(
+            cmd, capture_output=True, text=True, check=True
+        )
+        LOG.debug(
+            f"Command finished. stdout={process.stdout}, stderr={process.stderr}"
+        )
+        return Result(ResultType.COMPLETED)
+
+    def is_skip(self, status: Status | None = None) -> Result:
+        """Determines if the step should be skipped or not."""
+        machines_res = subprocess.run(
+            ["juju", "machines", "--format", "json"], capture_output=True
+        )
+        machines = json.loads(machines_res.stdout)["machines"]
+        n_machines = len(machines)
+        if n_machines > 2 and n_machines % 2 == 1:
+            machines_to_join = machines_missing_juju_controllers()
+            self.n = n_machines
+            self.extra_args.extend(("--to", ",".join(machines_to_join)))
+            LOG.debug(
+                f"Will enable Juju controller on machines {machines_to_join}"
+            )
+            return Result(ResultType.COMPLETED)
+        LOG.debug("Wrong number of machines, skipping scaling Juju")
+        return Result(ResultType.SKIPPED)
