+++ anvil-python/anvil/jobs/manifest.py
@@ -20,6 +20,7 @@ from pathlib import Path
 import shutil
 from typing import Any, Dict, List
 
+from pydantic import Field
 from pydantic.dataclasses import dataclass
 from snaphelpers import Snap
 from sunbeam import utils
@@ -36,7 +37,6 @@ from sunbeam.jobs.common import (
 from sunbeam.jobs.deployment import Deployment
 from sunbeam.jobs.manifest import (
     CharmsManifest,
-    JujuManifest,
     MissingTerraformInfoException,
     TerraformManifest,
 )
@@ -50,6 +50,19 @@ from anvil.versions import (
 )
 
 LOG = logging.getLogger(__name__)
+
+
+@dataclass
+class JujuManifest:
+    # Setting Field alias not supported in pydantic 1.10.0
+    # Old version of pydantic is used due to dependencies
+    # with older version of paramiko from python-libjuju
+    # Newer version of pydantic can be used once the below
+    # PR is released
+    # https://github.com/juju/python-libjuju/pull/1005
+    bootstrap_args: list[str] = Field(
+        default=[], description="Extra args for juju bootstrap"
+    )
 
 
 @dataclass(config=dict(extra="allow"))  # type: ignore[call-overload]
