+++ anvil-python/anvil/provider/local/commands.py
@@ -12,11 +12,8 @@
 # implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import asyncio
-import json
 import logging
 from pathlib import Path
-import subprocess
 from typing import List
 
 import click
@@ -76,7 +73,7 @@ from anvil.commands.haproxy import (
     RemoveHAProxyUnitStep,
     haproxy_install_steps,
 )
-from anvil.commands.juju import JujuAddSSHKeyStep
+from anvil.commands.juju import AnvilScaleJujuStep, JujuAddSSHKeyStep
 from anvil.commands.maas_agent import (
     RemoveMAASAgentUnitStep,
     maas_agent_install_steps,
@@ -96,10 +93,10 @@ from anvil.jobs.common import (
     roles_to_str_list,
     validate_roles,
 )
-from anvil.jobs.juju import CONTROLLER, AnvilScaleJujuStep
+from anvil.jobs.juju import CONTROLLER
 from anvil.jobs.manifest import Manifest
 from anvil.provider.local.deployment import LocalDeployment
-from anvil.utils import CatchGroup, machines_missing_juju_controllers
+from anvil.utils import CatchGroup
 
 LOG = logging.getLogger(__name__)
 console = Console()
@@ -501,24 +498,7 @@ def join(
                 name,
             )
         )
-
-    machines_res = subprocess.run(
-        ["juju", "machines", "--format", "json"], capture_output=True
-    )
-    machines = json.loads(machines_res.stdout)["machines"]
-    n_machines = len(machines)
-    if n_machines > 2 and n_machines % 2 == 1:
-        machines_to_join = machines_missing_juju_controllers()
-        LOG.debug(
-            f"Will enable Juju controller on machines {machines_to_join}"
-        )
-        plan2.append(
-            AnvilScaleJujuStep(
-                controller,
-                n_machines,
-                ["--to", ",".join(machines_to_join)],
-            )
-        )
+    plan2.append(AnvilScaleJujuStep(controller))
     run_plan(plan2, console)
 
     click.echo(f"Node joined cluster with roles: {pretty_roles}")
